# docs/openapi/v1.yaml
openapi: "3.0.3"
info:
  title: WebWaka OS API
  version: "1.0.1"
  description: |
    Multi-tenant, multi-vertical SaaS platform API for Africa (Nigeria-first).
    Governance-driven, edge-first on Cloudflare Workers.

    **Platform Invariants (non-negotiable):**
    - T3: Every query scoped by `tenant_id` from JWT — never from user-supplied input
    - P9: All monetary values in integer kobo — never floats or strings
    - P10: NDPR consent checked before any personal data collection
    - P13: No PII in AI requests — all AI paths through SuperAgent SDK

    **Authentication:** All protected routes require `Authorization: Bearer <jwt>` header.
    The JWT is issued by `POST /auth/login` and must include `userId`, `tenantId`, and `role`.
  contact:
    name: WebWaka Platform Team
    email: platform@webwaka.com
  license:
    name: Proprietary
    url: https://webwaka.com/terms
servers:
  - url: https://api.webwaka.com
    description: Production
  - url: https://api-staging.webwaka.com
    description: Staging
  - url: http://localhost:8787
    description: Local development (Wrangler dev)
...
paths:
...
  /fx-rates:
    get:
      tags: [FxRates]
      summary: List all active FX rates (public)
      operationId: listFxRates
      security: []
      responses:
        "200":
          description: Active FX rates (rate integer × 1,000,000 for float avoidance)
          content:
            application/json:
              schema:
                type: object
                properties:
                  rates:
                    type: array
                    items:
                      $ref: "#/components/schemas/FxRate"
                  total:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [FxRates]
      summary: Upsert an FX rate (super_admin only)
      operationId: upsertFxRate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [base_currency, quote_currency, rate, rate_inverse]
              properties:
                base_currency:
                  type: string
                  example: NGN
                quote_currency:
                  type: string
                  example: USD
                rate:
                  type: integer
                  description: Rate × 1,000,000 (e.g. 0.00065 NGN/USD = 650)
                  minimum: 1
                rate_inverse:
                  type: integer
                  description: Inverse rate × 1,000,000
                  minimum: 1
                source:
                  type: string
                  default: manual
                expires_in_hours:
                  type: integer
      responses:
        "201":
          description: FX rate created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /superagent/capabilities:
    get:
      tags: [SuperAgent]
      summary: List all AI capability types (public)
      description: |
        Returns the full catalogue of every `AICapabilityType` with human-readable
        `displayName`, `description`, `pillar` (1=Ops, 2=Branding, 3=Marketplace),
        and minimum `planTier` required.

        **No authentication required** — this is a public metadata endpoint used by
        workspace-apps and partner portals to drive capability-discovery UIs before
        a user session is established.

        Results are sorted by pillar (ascending) then alphabetically by capability key.
      operationId: listCapabilities
      security: []
      responses:
        "200":
          description: Full capability catalogue
          content:
            application/json:
              schema:
                type: object
                properties:
                  capabilities:
                    type: array
                    items:
                      type: object
                      properties:
                        capability:
                          type: string
                          description: The AICapabilityType key
                          example: inventory_ai
                        displayName:
                          type: string
                          example: Inventory Intelligence
                        description:
                          type: string
                        pillar:
                          type: integer
                          enum: [1, 2, 3]
                          description: "1=Ops, 2=Branding, 3=Marketplace"
                        planTier:
                          type: string
                          enum: [growth, pro, enterprise]

        "400":
          $ref: "#/components/responses/BadRequest"
...
  /superagent/compliance/check:
    get:
      tags: [SuperAgent]
      summary: Check compliance requirements for a vertical (SA-4.5)
      operationId: checkVerticalCompliance
      security:
        - BearerAuth: []
      parameters:
        - name: vertical
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Compliance profile for the vertical
          content:
            application/json:
              schema:
                type: object
                properties:
                  vertical:
                    type: string
                  is_sensitive:
                    type: boolean
                  sector:
                    type: string
                    nullable: true
                  requires_hitl:
                    type: boolean
                  hitl_level:
                    type: integer
                    nullable: true
                  disclaimers:
                    type: array
                    items:
                      type: string
        "400":
          description: vertical query parameter required
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
...
    FxRate:
      type: object
      required: [id, base_currency, quote_currency, rate, rate_inverse, source, effective_at]
      properties:
        id:
          type: string
        base_currency:
          type: string
          example: NGN
        quote_currency:
          type: string
          example: USD
        rate:
          type: integer
          description: Rate × 1,000,000 to avoid floats (actual = rate / 1_000_000)
        rate_inverse:
          type: integer
          description: Inverse rate × 1,000,000
        rate_actual:
          type: number
          description: Convenience field — rate / 1,000,000
        rate_inverse_actual:
          type: number
          description: Convenience field — rate_inverse / 1,000,000
        source:
          type: string
          enum: [manual, cbr, ecb, paystack]
        effective_at:
          type: integer
          description: Unix timestamp when rate becomes effective
        expires_at:
          type: integer
          nullable: true
          description: Unix timestamp when rate expires (null = no expiry)
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but insufficient role/permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"